+++ packages/project-template-browser-react/src/containers/HomePage/HomePage.tsx
@@ -1,0 +1,5 @@
+import { Introduction } from 'src/components/Introduction';
+
+export function HomePage(): React.ReactElement {
+    return <Introduction />;
+}